solved some build warnings and unnecessary error logs
